+++ gnsq/stream/stream.py
@@ -2,7 +2,7 @@
 from __future__ import absolute_import
 
 from resource import getpagesize
-from errno import ENOTCONN
+from errno import ENOTCONN, EDEADLK, EAGAIN, EWOULDBLOCK
 
 import gevent
 from gevent import socket
@@ -70,6 +70,9 @@ class Stream(object):
             try:
                 packet = self.socket.recv(self.buffer_size)
             except socket.error as error:
+                if error.errno in (EDEADLK, EAGAIN, EWOULDBLOCK):
+                    gevent.sleep()
+                    continue
                 raise NSQSocketError(*error)
 
             if not packet:
@@ -104,7 +107,6 @@ class Stream(object):
 
         self.state = DISCONNECTED
         self.queue.put(StopIteration)
-        self.socket.close()
 
     def send_loop(self):
         for data, result in self.queue:
@@ -121,6 +123,8 @@ class Stream(object):
 
             except Exception as error:
                 result.set_exception(error)
+
+        self.socket.close()
 
     def upgrade_to_tls(
         self,