+++ tests/conftest.py
@@ -3,12 +3,12 @@ import pytest
 
 def pytest_addoption(parser):
     parser.addoption(
-        '--runslow',
+        '--fast',
         action='store_true',
-        help='run slow tests'
+        help='do not run slow tests'
     )
 
 
 def pytest_runtest_setup(item):
-    if 'slow' in item.keywords and not item.config.getoption('--runslow'):
-        pytest.skip('need --runslow option to run')
+    if 'slow' in item.keywords and item.config.getoption('--fast'):
+        pytest.skip('skiping because of --fast')