+++ HISTORY.rst
@@ -3,6 +3,14 @@
 History
 -------
 
+0.1.1 (2014-07-07)
+~~~~~~~~~~~~~~~~~~
+
+* Fix packaging stream submodule.
+* Send queued messages before closing socket.
+* Continue to read from socket on EAGAIN
+
+
 0.1.0 (2014-07-07)
 ~~~~~~~~~~~~~~~~~~